project full authentication and authorization

+++ backend/src/main/java/se/example2/softhouse/DAO/ExamDAO.java
@@ -27,16 +27,16 @@ public interface ExamDAO {
 
     @GetGeneratedKeys
     @SqlUpdate("insert into Exam (id, text) values (:id, :text)")
-    int create(@BindBean Exam exam);
+    long create(@BindBean Exam exam);
 
     @SqlQuery("select * from Exam where id = :id")
-    Exam retrieve(@Bind("id") int id);
+    Exam retrieve(@Bind("id") long id);
 
     @SqlUpdate("delete from Exam where id = :id")
-    void delete(@Bind("id") int id);
+    void delete(@Bind("id") long id);
 
     @SqlUpdate("update Exam set text = :u.text where id = :id")
-    void update(@Bind("id") int id, @BindBean("u") Exam exam);
+    void update(@Bind("id") long id, @BindBean("u") Exam exam);
 
     @SqlQuery("select * from Exam where text = :text")
     Exam retrieveByText(@Bind("text") String text);
